Revert "add simple test"
testing
