+++ myapplication.py
@@ -1,0 +1,56 @@
+'''
+Simple Flask application to test deployment to Amazon Web Services
+Uses Elastic Beanstalk and RDS
+
+Author: Scott Rodkey - rodkeyscott@gmail.com
+
+Step-by-step tutorial: https://medium.com/@rodkey/deploying-a-flask-application-on-aws-a72daba6bb80
+'''
+
+import logging
+from flask import render_template, request
+from application import db, application
+from application.models import Data
+from application.forms import EnterDBInfo, RetrieveDBInfo
+from environs import Env
+
+env = Env()
+env.read_env()
+
+logging.basicConfig(level=logging.INFO)
+
+db.create_all()
+
+@application.route('/', methods=['GET', 'POST'])
+@application.route('/index', methods=['GET', 'POST'])
+def index():
+    form1 = EnterDBInfo(request.form) 
+    form2 = RetrieveDBInfo(request.form) 
+    
+    if request.method == 'POST' and form1.validate():
+        data_entered = Data(notes=form1.dbNotes.data)
+        try:     
+            db.session.add(data_entered)
+            db.session.commit()        
+            db.session.close()
+        except:
+            db.session.rollback()
+        return render_template('thanks.html', notes=form1.dbNotes.data)
+        
+    if request.method == 'POST' and form2.validate():
+        try:   
+            num_return = int(form2.numRetrieve.data)
+            query_db = Data.query.order_by(Data.id.desc()).limit(num_return)
+            for q in query_db:
+                print(q.notes)
+            db.session.close()
+        except:
+            db.session.rollback()
+        return render_template('results.html', results=query_db, num_return=num_return)                
+    
+    return render_template('index.html', form1=form1, form2=form2)
+
+if __name__ == '__main__':
+    application.debug = True
+    application.run()
+    logging.info('Application run command complete.')
